05/06

+++ book-shop/controller/OrderController.js
@@ -3,6 +3,8 @@ const db = require('mysql2/promise');
 const {StatusCodes} = require('http-status-codes');
 const dotenv = require('dotenv');
 dotenv.config();
+const {ensureAuthorization} = require('../auth');
+const jwt = require('jsonwebtoken');
 
 
 const order = async (req, res) => {
@@ -15,49 +17,71 @@ const order = async (req, res) => {
         dateStrings : true
     });
 
-    const {items, delivery, totalQuantity, totalPrice, userId, firstBookTitle} = req.body;
+    let authorization = ensureAuthorization(req, res);
 
+    if (authorization instanceof jwt.TokenExpiredError) {
+        return res.status(StatusCodes.UNAUTHORIZED).json({message : "Token expired"});
+    }
+    else if (authorization instanceof jwt.JsonWebTokenError) {
+        return res.status(StatusCodes.UNAUTHORIZED).json({message : "Invalid token"});
+    }
+    else {
+        const {items, delivery, totalQuantity, totalPrice, firstBookTitle} = req.body;
 
-    let sql = `INSERT INTO delivery (address, receiver, contact) VALUES (?, ?, ?);`
-    let values = [delivery.address, delivery.receiver, delivery.contact];
+        //delivery 테이블 삽입
+        let sql = `INSERT INTO delivery (address, receiver, contact) VALUES (?, ?, ?);`
+        let values = [delivery.address, delivery.receiver, delivery.contact];
 
-    let [results] = await conn.execute(sql, values);
+        let [results] = await conn.execute(sql, values);
 
-    let delivery_id = results.insertId;
+        let delivery_id = results.insertId;
 
-    //order 테이블 삽입
-    sql = `INSERT INTO orders (book_title, total_quantity, total_price, user_id, delivery_id) 
-            VALUES (?, ?, ?, ?, ?);`;
-    values = [firstBookTitle, totalQuantity, totalPrice, userId, delivery_id];
-    [results] = await conn.execute(sql, values);
+        //order 테이블 삽입
+        sql = `INSERT INTO orders (book_title, total_quantity, total_price, user_id, delivery_id) 
+                VALUES (?, ?, ?, ?, ?);`;
+        values = [firstBookTitle, totalQuantity, totalPrice, authorization.id, delivery_id];
+        [results] = await conn.execute(sql, values);
 
-    let order_id = results.insertId;
+        let order_id = results.insertId;
 
-    // items 으로 -> book_id, quantity 조회
-    sql = `SELECT book_id, quantity FROM cartItems WHERE IN (?)`;
-    let [orderItems, fields] = await conn.query(sql, [items]);
+        // items 으로 -> book_id, quantity 조회
+        sql = `SELECT book_id, quantity FROM cartItems WHERE id IN (?)`;
+        let [orderItems, fields] = await conn.query(sql, [items]);
 
-    //orderedBook 테이블 삽입
-    sql = `INSERT INTO orderedBook (order_id, book_id, quantity) VALUES ?`;
+        //orderedBook 테이블 삽입
+        sql = `INSERT INTO orderedBook (order_id, book_id, quantity) VALUES ?`;
 
-    values = [];
-    orderItems.forEach((item) => {
-        values.push([order_id, item.book_id, item.quantity]);
-    });
-    results = await conn.query(sql, [values]);
+        values = [];
+        orderItems.forEach((item) => {
+            values.push([order_id, item.book_id, item.quantity]);
+        });
+        results = await conn.query(sql, [values]);
 
-    let del = await deleteCartItems(conn, items);
+        let del = await deleteCartItems(conn, items);
 
-    return res.status(StatusCodes.OK).json(results[0]);
+        return res.status(StatusCodes.OK).json(results[0]);
 
+    }
+   
 };
 
 const deleteCartItems = async (conn, items) => {
     let sql = "DELETE FROM cartItems WHERE id IN (?);";
 
-    let results = await conn.query(sql, [items]);
+let results = await conn.query(sql, [items]);
     return results;
 }
+
+const toCamel = (s) =>
+    s.replace(/_([a-z])/g, (g) => g[1].toUpperCase());
+  
+  const camelizeRow = (row) => {
+    const newRow = {};
+    for (let key in row) {
+      newRow[toCamel(key)] = row[key];
+    }
+    return newRow;
+  };
 
 
 const getOrders = async (req, res) => {
@@ -76,13 +100,14 @@ const getOrders = async (req, res) => {
                 ON orders.delivery_id = delivery.id;`;
 
     let [rows, fields] = await conn.query(sql);
-    return res.status(StatusCodes.OK).json(rows);
+    const camelRows = rows.map(camelizeRow);
+    return res.status(StatusCodes.OK).json(camelRows);
 
 };
 
 const getOrderDetail = async (req, res) => {
 
-    const {id} = req.params;
+    const orderId = req.params.id;
 
     const conn = await db.createConnection({
         host : process.env.DB_HOST,
@@ -98,8 +123,9 @@ const getOrderDetail = async (req, res) => {
                 ON orderedBook.book_id = books.id
                 WHERE order_id = ?`;
 
-    let [rows, fields] = await conn.query(sql, [id]);
-    return res.status(StatusCodes.OK).json(rows);
+    let [rows, fields] = await conn.query(sql, [orderId]);
+    const camelRows = rows.map(camelizeRow);
+    return res.status(StatusCodes.OK).json(camelRows);
 
 };
